requirements.txt
asgiref==3.3.1
attrs==20.3.0
Automat==20.2.0
cffi==1.14.4
constantly==15.1.0
cryptography==3.2.1
cssselect==1.1.0
Django==3.1.4
django-environ==0.4.5
django-phone-field==1.8.1
djangorestframework==3.12.2
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
itemadapter==0.2.0
itemloaders==1.0.4
jmespath==0.10.0
lxml==4.6.2
parsel==1.6.0
Protego==0.1.16
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pyOpenSSL==20.0.0
pytz==2020.4
queuelib==1.5.0
Scrapy==2.4.1
scrapy-djangoitem==1.1.1
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
Twisted==20.3.0
w3lib==1.22.0
yapf==0.30.0
zope.interface==5.2.0
